requirements.txt
aiohttp==3.8.3
aiomysql==0.1.1
aiosignal==1.2.0
aniso8601==9.0.1
anyio==3.6.1
arrow==1.2.3
async-timeout==4.0.2
attrs==22.1.0
Babel==2.10.3
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colour==0.1.5
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
frozenlist==1.3.1
greenlet==1.1.2
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
lxml==4.9.1
MarkupSafe==2.1.1
multidict==6.0.2
orjson==3.8.0
pip==21.3.1
playwright-stealth==1.0.5
playwright==1.25.2
protobuf==3.20.1
pydantic==1.10.2
pyee==8.1.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.2.1
pyuseragents==1.0.5
PyYAML==6.0
requests-toolbelt==0.9.1
requests==2.28.1
setuptools==60.2.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
speaklater==1.3
SQLAlchemy-Utils==0.38.3
SQLAlchemy==1.4.41
starlette==0.20.4
tablib==3.2.1
typing-extensions==4.4.0
ujson==5.5.0
urllib3==1.26.9
uvicorn==0.18.3
uvloop==0.17.0
cryptography==38.0.1
